requirements.txt
vk_api==11.9.6
requests==2.25.0
python-telegram-bot==13.7
python-dotenv==0.19.2
google-cloud-dialogflow==2.9.1

